frontend-app/requirements.txt
Flask==2.3.3
google-cloud-storage==1.29.0
gunicorn==21.2.0
